requirements.txt
Django==2.2.13
pytz==2018.5
djangorestframework==3.14.0
ujson==5.4.0
pyyaml==6.0